Cambio de nombre del paquete, implementacion de de actividad de usuario primera vez y limpieza de codigo.
