+++ app/src/main/java/com/example/cookplus/BuscarRecetasAdapter.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.cookplus;
 
 import android.app.Activity;
 import android.content.Context;
@@ -10,7 +10,6 @@ import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import java.util.Arrays;
 import java.util.List;
 
 /**
