+++ app/src/main/java/com/example/cookplus/DbAdapter.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.cookplus;
 
 import android.content.ContentValues;
 import android.content.Context;
