+++ app/src/main/java/com/example/cookplus/FavItem.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.cookplus;
 
 class FavItem {
        private String titulo;
