+++ app/src/main/java/com/example/cookplus/MainActivity.java
@@ -1,13 +1,9 @@
-package com.example.myapplication;
+package com.example.cookplus;
 
 import androidx.appcompat.app.AppCompatActivity;
-import android.util.Log;
+
 import android.os.Bundle;
 import android.content.Intent;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.content.Intent;
-import android.os.Bundle;
 import android.os.Handler;
