+++ app/src/main/java/com/example/cookplus/NeveraActivity.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.cookplus;
 
 
 import android.app.AlertDialog;
