+++ app/src/main/java/com/example/cookplus/NeveraItem.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.cookplus;
 
 public class NeveraItem {
     private String text;
