+++ app/src/main/java/com/example/cookplus/Receta.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.cookplus;
 
 import android.graphics.Bitmap;
