+++ app/src/main/res/menu/menu_usuario.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.myapplication.UsuarioActivity">
+    tools:context="com.example.cookplus.UsuarioActivity">
 
     <item
         android:id="@+id/action_edit"
