+++ app/src/main/res/menu/navigation_bottom.xml
@@ -1,9 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <!-- <item
-         android:id="@+id/action_back"
-         android:icon="@drawable/arrow_back"
-         android:title="@string/menu_back" /> -->
     <item
         android:id="@+id/action_favorite"
         android:icon="@drawable/fav"
