+++ app/src/test/java/com/example/cookplus/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.cookplus;
 
 import org.junit.Test;
